rename endpoints and endpoint methods, move recognition loop to separate module, clean code a little
