+++ src/main.rs
@@ -16,17 +16,20 @@ extern crate tch;
 
 mod helpers;
 mod neuro_net;
+mod recognition;
 mod router;
 mod settings;
 
 use base64::{decode, encode};
 use crossbeam_channel::{unbounded, Receiver, Sender};
-use helpers::{delete_file, save_file};
+use helpers::{delete_file, save_file, stop_program};
 use neuro_net::{draw_results, report, Bbox};
+use recognition::run_recognition_ai;
 use rocket_contrib::json::Json;
 use router::{ImageRequestVectorized, RequestType};
 use settings::{
-    CONFIDENCE_THRESHOLD, CONFIG, IS_CUDA, NAMES, NMS_THRESHOLD, TEMPORARY_FILE_PATH, WEIGHTS,
+    CONFIDENCE_THRESHOLD, CONFIG, CUDA_THREADS, IS_CUDA, NAMES, NMS_THRESHOLD, TEMPORARY_FILE_PATH,
+    WEIGHTS,
 };
 use std::collections::BTreeMap;
 use std::fs::remove_file;
@@ -53,83 +56,23 @@ fn main() {
     let (sender_imgage_bboxed, receiver_image_bboxed): (Sender<Vec<u8>>, Receiver<Vec<u8>>) =
         unbounded();
     thread::spawn(move || {
-        let mut var_store: VarStore = VarStore::new(tch::Device::Cpu);
-        if IS_CUDA {
-            var_store = VarStore::new(tch::Device::Cuda(16));
-        }
-        if let Ok(darknet) = neuro_net::parse_config(&CONFIG.to_string()) {
-            let net_width = darknet.width().expect("Cannot get darknet width");
-            let net_height = darknet.height().expect("Cannot get darknet height");
-            if let Ok(model) = darknet.build_model(&var_store.root()) {
-                if let Err(_) = var_store.load(&WEIGHTS.to_string()) {
-                    stop_program();
-                };
-                loop {
-                    if let Ok(img_request) = receiver_img.recv() {
-                        if let Ok(_) = save_file(&img_request.img, &TEMPORARY_FILE_PATH.to_string())
-                        {
-                            let original_image =
-                                image_helper::load(&TEMPORARY_FILE_PATH.to_string())
-                                    .expect("Cannot load image from file");
-                            let image_resized: Tensor =
-                                image_helper::resize(&original_image, net_width, net_height)
-                                    .expect("Cannot resize image");
-                            let image_calc_ready: Tensor =
-                                image_resized.unsqueeze(0).to_kind(tch::Kind::Float) / 255.;
-                            let predictions = model.forward_t(&image_calc_ready, false).squeeze();
-                            let bboxes =
-                                report(&predictions, &original_image, net_width, net_height)
-                                    .expect("Cannot generate report");
-                            match img_request.request {
-                                RequestType::BBOXES => {
-                                    if let Err(_) = sender_plane_bboxed.send(bboxes) {
-                                        error!("Cannot send bboxes solution between threads.");
-                                    }
-                                }
-                                RequestType::IMAGE => {
-                                    if let Ok(img_drawed_boxes) =
-                                        draw_results(&image_resized, bboxes)
-                                    {
-                                        let image_buffer: Vec<u8> = Vec::from(img_drawed_boxes);
-                                        println!(
-                                            "Tensor to image buffer of length {:?}",
-                                            &image_buffer.len()
-                                        );
-                                        if let Err(_) = sender_imgage_bboxed.send(image_buffer) {
-                                            error!("Cannot send image buffer solution between threads.");
-                                        }
-                                    };
-                                }
-                            };
-                            if let Err(_) = delete_file(&TEMPORARY_FILE_PATH.to_string()) {
-                                panic!("It is not possible to delete saved temporary file, something went terribly wrong.");
-                            }
-                        } else {
-                            panic!("It is not possible to save temporary file, something went terribly wrong.");
-                        }
-                    }
-                }
-            } else {
-                stop_program();
-            }
-        } else {
-            stop_program();
-        }
+        run_recognition_ai(
+            receiver_img.clone(),
+            sender_plane_bboxed.clone(),
+            sender_imgage_bboxed.clone(),
+        )
     });
     rocket::ignite()
         .manage(Arc::new(sender_img.clone()))
         .manage(Arc::new(receiver_plane_bboxed.clone()))
         .manage(Arc::new(receiver_image_bboxed.clone()))
         .mount(
-            "/mirror",
-            routes![router::tell_me_who, router::show_me_who, router::get_names],
+            "/recognition/object",
+            routes![
+                router::get_names,
+                router::post_recognize_objects_boxes,
+                router::post_recognize_objects_images
+            ],
         )
         .launch();
-}
-
-fn stop_program() {
-    panic!(format!(
-        "Fatal error, darknet cannot be loaded, or some other dependencies are not satisfied, 
-        please look in to documentation to check if all steps has been taken."
-    ));
 }
